Revert "FastAccess improvements: handle edge cases, add StructSetterHandler", remove recently added FastAccess unit tests, rename TestFastAccess to TestFieldRefAccess
